Setup (#250)
* allow hdf5:// prefix - #247

* fix flake8 errors

* use tcp by default in hsds app

* update test to run with quick stat

* add toml file

* catch login name not defined

* remove defunct rangeget test

* updated readme quick start

* bump aiohttp version to 3.8.5

* fix quick start instructions

+++ build.sh
@@ -27,11 +27,17 @@ if [ $run_pyflakes ]; then
     fi
 fi
 
-echo "running setup.py"
-python setup.py install
+pip install --upgrade build
 
-echo "clean stopped containers"
-docker rm -v $(docker ps -aq -f status=exited)
+echo "running build"
+python -m build
+pip install -v .
 
-echo "building docker image"
-docker build -t hdfgroup/hsds .
+command -v docker 
+if [ $? -ne 1 ]; then
+    echo "clean stopped containers"
+    docker rm -v $(docker ps -aq -f status=exited)
+
+    echo "building docker image"
+    docker build -t hdfgroup/hsds .
+fi
